// src/app/model/user-info.model.ts
export class UserInfo {
    age?: number;
    avatarData?: string;
    name?: string;
    studentId?: number;
    teacherId?: number;
    userId: number;
    userType: string;
    username: string;
    expPoints?: number;
    studentReadingLevel?: string;
    quizSession?: object;
    teacher?: object;
    gender: string;
    avatarType?: string;
    userAvatarId?: string;
}